requirements.txt
openai==1.40.0
httpx==0.27.0
flask
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client